+++ lib/ioc.dart
@@ -5,3 +5,6 @@ export 'src/ioc/application_host.dart';
 export 'src/ioc/base_service.dart';
 export 'src/ioc/service_host.dart';
 export 'src/ioc/service_resolver.dart';
+export 'src/ioc/isolated_host.dart';
+export 'src/ioc/isolated_host_configuration.dart';
+export 'src/ioc/isolated_service.dart';