+++ lib/src/ioc/isolated_host_configuration.dart
@@ -1,0 +1,11 @@
+import 'package:datahub/services.dart';
+
+class IsolatedHostConfiguration {
+  final Map<String, dynamic> config;
+  final LogBackend logBackend;
+
+  IsolatedHostConfiguration(
+    this.config,
+    this.logBackend,
+  );
+}